+++ src/main/resources/clojure/install/ClojureTools.psm1
@@ -173,15 +173,15 @@ in a terminal, and should be preferred unless you don't want that support,
 then use 'clojure'.
 
 Usage:
-  Start a REPL   clj     [clj-opt*] [-A:aliases] [init-opt*]
-  Exec function  clojure [clj-opt*] -X[:aliases] [a/fn] [kpath v]*
-  Run main       clojure [clj-opt*] -M[:aliases] [init-opt*] [main-opt] [arg*]
+  Start a REPL   clj     [clj-opt*] [-Aaliases] [init-opt*]
+  Exec function  clojure [clj-opt*] -X[aliases] [a/fn] [kpath v]*
+  Run main       clojure [clj-opt*] -M[aliases] [init-opt*] [main-opt] [arg*]
   Prepare        clojure [clj-opt*] -P [other exec opts]
 
 exec-opts:
-  -A:aliases     Use aliases to modify classpath
-  -X[:aliases]   Use aliases to modify classpath or supply exec fn/args
-  -M[:aliases]   Use aliases to modify classpath or supply main opts
+  -Aaliases      Use concatenated aliases to modify classpath
+  -X[aliases]    Use concatenated aliases to modify classpath or supply exec fn/args
+  -M[aliases]    Use concatenated aliases to modify classpath or supply main opts
   -P             Prepare deps - download libs, cache classpath, but don't exec
 
 clj-opts: